+++ crates/bevy_input/src/mouse/mod.rs
@@ -1,6 +1,6 @@
 pub mod button;
 pub mod event;
+pub mod scrolling;
 pub mod system;
-pub mod unit;
 
-pub use crate::mouse::{button::*, event::*, system::*, unit::*};
+pub use crate::mouse::{button::*, event::*, scrolling::*, system::*};
